Truncation warnings fixes for AssetTools
#rb Devin.Doucette
#jira UE-166274

[CL 26758160 by Francis Hurteau in 5.3 branch]

+++ Engine/Source/Developer/AssetTools/Private/AssetFixUpRedirectors.cpp
@@ -208,7 +208,7 @@ bool FAssetFixUpRedirectors::UpdatePackageStatus(const TArray<FRedirectorRefs>& 
 
 void FAssetFixUpRedirectors::LoadReferencingPackages(TArray<FRedirectorRefs>& RedirectorsToFix, TArray<UPackage*>& OutReferencingPackagesToSave, TArray<UPackage*>& OutLoadedPackages) const
 {
-	FScopedSlowTask SlowTask( RedirectorsToFix.Num(), LOCTEXT( "LoadingReferencingPackages", "Loading Referencing Packages..." ) );
+	FScopedSlowTask SlowTask( static_cast<float>(RedirectorsToFix.Num()), LOCTEXT( "LoadingReferencingPackages", "Loading Referencing Packages..." ) );
 	SlowTask.MakeDialog();
 
 	ISourceControlProvider& SourceControlProvider = ISourceControlModule::Get().GetProvider();
